https://kukuta.tistory.com/448 - 랜덤 던전 3D 모델 적용

+++ 448/Assets/Scripts/NDungeon/Tile.cs
@@ -15,7 +15,7 @@ namespace NDungeon
             get => data.type;
         }
 
-        public Tile[] neighbors = new Tile[(int)NDungeon.NTileMap.TileMap.Tile.Direction.Max];
+        public Tile[] neighbors = new Tile[NDungeon.NTileMap.TileMap.Tile.Direction.Max];
         public NTileMap.TileMap.Tile data;
     }
 }
